Explicitly check if file is a directory in merge_xml_files
PHP 7.4 and above emit E_WARNING if file_get_contents() is called with a
directory as the argument, so in case of a subdirectory this warning
was logged but the subfiles were not added to the array. Explicitly
check if $file is a directory before calling file_get_contents() to make
this work with PHP 7.4 and above.